Move TypeName conversion to jet module
